+++ webcomponents/core/src/core.js
@@ -3,6 +3,35 @@ import { trivetCSS } from '@trvt/assets';
 
 import { createFragment } from './createFragment.js';
 
+const colorScheme = {
+	dark: window.matchMedia('(prefers-color-scheme: dark)'),
+	light: window.matchMedia('(prefers-color-scheme: light)'),
+	_active: undefined,
+	get current() {
+		return colorScheme._active || colorScheme.light.matches ? 'light' : 'dark';
+	},
+	set current(mode) {
+		colorScheme._active = mode;
+	},
+};
+colorScheme.dark.addEventListener(
+	'change',
+	e => e.matches && activateColorScheme({ mode: 'dark' })
+);
+colorScheme.light.addEventListener(
+	'change',
+	e => e.matches && activateColorScheme({ mode: 'light' })
+);
+
+function activateColorScheme({ mode }) {
+	if (
+		!document.documentElement.dataset.colorScheme ||
+		colorScheme.current !== mode
+	) {
+		document.documentElement.dataset.colorScheme = mode;
+	}
+}
+
 /**
  * cssLayerDefinitions
  * @type {Array}
@@ -23,6 +52,7 @@ class TrivetElement extends HTMLElement {
 	#template;
 
 	#render() {
+		activateColorScheme({ mode: colorScheme.current });
 		this.#shadowRoot.innerHTML = this.template;
 	}